refactor: split octree/cam_core tests and replace tolerance magic numbers
